requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
openai==1.99.9
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.30
chromadb==1.0.17